requirements.txt
torch==2.10.0
pyyaml==6.0.3
matplotlib==3.10.8
numpy==2.4.2
tqdm==4.67.3
sentencepiece==0.2.1